requirements.txt
Django==2.0
djangorestframework==3.8
flake8
coverage
coveralls
psycopg2-binary
django-cors-headers
mock
